feat(bono-tecnicos): soportar medios dias en el campo Dias
El Excel legacy admite medios días (caso real: Matías Milán, bono
julio/2026, 20,5 días) pero la columna era Integer NOT NULL. Se migró a
Numeric(4,1) en Postgres, dominio/DTOs/schemas a float con validación de
múltiplos de 0.5, y el input de la tabla (EditableNumberCell) a step=0.5
en vez de truncar a entero.

+++ backend/src/modules/bono_tecnicos/application/dtos/puntaje_tecnico_dto.py
@@ -16,7 +16,7 @@ class PuntajeTecnicoDTO:
     inst_des: int
     pre_correctivo: int
     entrega_insumos: int
-    dias: int
+    dias: float
     tareas_varias: int
     puntaje: float | None
     # None si el técnico no está vinculado a un empleado de Gestión de
@@ -29,4 +29,4 @@ class GuardarBonoInputRequest:
     id_tecnico: int
     periodo: int
     tecnico: str
-    dias: int
+    dias: float
